1.add new job for dianping count   2. change dianping variable to long type

+++ modules/quhao.appserver/app/com/withiter/models/account/AccountEntityDef.java
@@ -33,7 +33,7 @@ public abstract class AccountEntityDef extends BaseModel {
 
 	public int jifen = 5;
 	
-	public int dianping = 0;
+	public long dianping = 0;
 	public int zhaopian = 0;
 	
 	public String birthDay = "";
